Split FunctionsStringArray
