+++ src/Functions/URL/URLPathHierarchy.cpp
@@ -1,8 +1,13 @@
 #include <Functions/FunctionFactory.h>
-#include <Functions/FunctionsStringArray.h>
+#include <Functions/FunctionTokens.h>
 
 namespace DB
 {
+
+namespace
+{
+
+using Pos = const char *;
 
 class URLHierarchyImpl
 {
@@ -13,7 +18,6 @@ private:
 
 public:
     static constexpr auto name = "URLHierarchy";
-    static String getName() { return name; }
 
     static bool isVariadic() { return false; }
     static size_t getNumberOfArguments() { return 1; }
@@ -97,8 +101,9 @@ public:
 };
 
 
-struct NameURLHierarchy { static constexpr auto name = "URLHierarchy"; };
 using FunctionURLHierarchy = FunctionTokens<URLHierarchyImpl>;
+
+}
 
 REGISTER_FUNCTION(URLHierarchy)
 {
