+++ src/Functions/URL/extractURLParameterNames.cpp
@@ -1,8 +1,13 @@
 #include <Functions/FunctionFactory.h>
-#include <Functions/FunctionsStringArray.h>
+#include <Functions/FunctionTokens.h>
 
 namespace DB
 {
+
+namespace
+{
+
+using Pos = const char *;
 
 class ExtractURLParameterNamesImpl
 {
@@ -13,7 +18,6 @@ private:
 
 public:
     static constexpr auto name = "extractURLParameterNames";
-    static String getName() { return name; }
 
     static bool isVariadic() { return false; }
     static size_t getNumberOfArguments() { return 1; }
@@ -80,8 +84,9 @@ public:
     }
 };
 
-struct NameExtractURLParameterNames { static constexpr auto name = "extractURLParameterNames"; };
 using FunctionExtractURLParameterNames = FunctionTokens<ExtractURLParameterNamesImpl>;
+
+}
 
 REGISTER_FUNCTION(ExtractURLParameterNames)
 {
