+++ src/Functions/URL/extractURLParameters.cpp
@@ -1,8 +1,14 @@
 #include <Functions/FunctionFactory.h>
-#include <Functions/FunctionsStringArray.h>
+#include <Functions/FunctionTokens.h>
+
 
 namespace DB
 {
+
+namespace
+{
+
+using Pos = const char *;
 
 class ExtractURLParametersImpl
 {
@@ -13,7 +19,6 @@ private:
 
 public:
     static constexpr auto name = "extractURLParameters";
-    static String getName() { return name; }
 
     static bool isVariadic() { return false; }
     static size_t getNumberOfArguments() { return 1; }
@@ -88,8 +93,9 @@ public:
     }
 };
 
-struct NameExtractURLParameters { static constexpr auto name = "extractURLParameters"; };
 using FunctionExtractURLParameters = FunctionTokens<ExtractURLParametersImpl>;
+
+}
 
 REGISTER_FUNCTION(ExtractURLParameters)
 {
